+++ src/backend/dialog/ipc_handler.ts
@@ -1,0 +1,10 @@
+import { ipcMain } from 'electron'
+import { showErrorBoxModalAuto } from './utils'
+
+ipcMain.handle(
+  'showErrorBox',
+  async (event, args: [title: string, error: string]) => {
+    const [title, error] = args
+    return showErrorBoxModalAuto({ event, title, error })
+  }
+)
