+++ src/backend/launcher.ts
@@ -16,7 +16,6 @@ import {
   getSteamRuntime,
   isEpicServiceOffline,
   isOnline,
-  showErrorBoxModalAuto,
   searchForExecutableOnPath,
   quoteIfNecessary,
   errorHandler,
@@ -49,6 +48,7 @@ import {
 import { spawn } from 'child_process'
 import shlex from 'shlex'
 import { Game } from './games'
+import { showErrorBoxModalAuto } from './dialog/utils'
 
 async function prepareLaunch(
   game: LegendaryGame | GOGGame,
