+++ src/backend/shortcuts/nonesteamgame/nonesteamgame.ts
@@ -10,7 +10,7 @@ import { join } from 'path'
 import { GameInfo } from 'common/types'
 import { ShortcutsResult } from '../types'
 import { getIcon } from '../utils'
-import { notify, showErrorBoxModalAuto } from '../../utils'
+import { notify } from '../../utils'
 import {
   prepareImagesForSteam,
   generateShortcutId,
@@ -22,6 +22,7 @@ import { app } from 'electron'
 import { isFlatpak, isWindows, tsStore } from '../../constants'
 import { logError, logInfo, LogPrefix, logWarning } from '../../logger/logger'
 import i18next from 'i18next'
+import { showErrorBoxModalAuto } from '../../dialog/utils'
 
 /**
  * Opens a error dialog in frontend with the error message
