+++ src/backend/tools.ts
@@ -3,16 +3,12 @@ import * as axios from 'axios'
 import { existsSync, readFileSync, writeFileSync } from 'graceful-fs'
 import { exec, spawn } from 'child_process'
 
-import {
-  execAsync,
-  getWineFromProton,
-  isOnline,
-  showErrorBoxModalAuto
-} from './utils'
+import { execAsync, getWineFromProton, isOnline } from './utils'
 import { execOptions, heroicToolsPath, userHome } from './constants'
 import { logError, logInfo, LogPrefix, logWarning } from './logger/logger'
 import i18next from 'i18next'
 import { dirname } from 'path'
+import { showErrorBoxModalAuto } from './dialog/utils'
 
 export const DXVK = {
   getLatest: async () => {
